data: withDirectoryFromName: move to helpers
Also remove the more-composable default behavior of using the
current thing's name, because this is no longer meant to be
composed generally.
